# lite-backend/llm-config-gateway/openapi.yaml
openapi: 3.0.3
info:
  title: LLM Config Gateway
  version: 0.1.0
servers:
  - url: http://localhost:9050
paths:
...
  /v1/defaults/simple:
    get:
      summary: Get simplified defaults (name + provider)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  chat:
                    type: object
                    properties:
                      model: { type: string }
                      provider: { type: string }
                  embedding:
                    type: object
                    properties:
                      model: { type: string }
                      provider: { type: string }
                  rerank:
                    type: object
                    properties:
                      model: { type: string }
                      provider: { type: string }
    post:
      summary: Set defaults
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                default_model:
                  type: string
                  description: model id or alias; empty string clears; literal "string" is ignored
                default_embedding:
                  type: string
                  description: embedding model id or alias; empty string clears; literal "string" is ignored
                default_rerank:
                  type: string
                  description: rerank model id or alias; empty string clears; literal "string" is ignored
                provider:
                  type: string
                  description: optional provider name/type when setting default_model to help accurate mapping
                provider_id:
                  type: integer
                  description: optional provider id when setting default_model
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                allOf:
                  - $ref: '#/components/schemas/DefaultsResponse'
...
  /v1/chat/completions:
    post:
      summary: OpenAI compatible chat completions (basic)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                model: { type: string }
                messages:
                  type: array
                  items:
                    type: object
                stream: { type: boolean }
      responses:
        '200': { description: OK }
...
  /mcp/registry/{tenant}/{name}/tools:
    get:
      tags: [MCP]
      summary: Get cached tools for MCP instance
      parameters:
        - in: path
          name: tenant
          required: true
          schema: { type: string }
        - in: path
          name: name
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MCPTool'
    put:
      tags: [MCP]
      summary: Update cached tools for MCP instance
      parameters:
        - in: path
          name: tenant
          required: true
          schema: { type: string }
        - in: path
          name: name
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MCPUpdateToolsRequest'
      responses:
        '200': { description: OK }
...
  /mcp/servers/{name}/tools:
    get:
      tags: [MCP]
      summary: List tools for a running server (live if available, else cached)
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MCPTool'
  /mcp/servers/{name}/tools/{tool}/call:
    post:
      tags: [MCP]
      summary: Call a tool via shared runtime (stdio)
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
        - in: path
          name: tool
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200': { description: OK }
...
  /api-tools/configs/{name}/tools:
    get:
      tags: [API-Tools]
      summary: List tools under given config
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
...
components:
  schemas:
...
    MCPTool:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
...
    MCPUpdateToolsRequest:
      type: object
      required: [tools]
      properties:
        tools:
          type: array
          items:
            $ref: '#/components/schemas/MCPTool'
    DefaultsResponse:
      type: object
      properties:
        default_model: { type: string }
        default_embedding: { type: string }
        default_model_info:
          $ref: '#/components/schemas/DefaultModelInfo'
        default_embedding_info:
          $ref: '#/components/schemas/DefaultModelInfo'
    ProviderInfo:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        type: { type: string }
        base_url: { type: string }
        status: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    DefaultModelInfo:
      type: object
      properties:
        id: { type: integer }
        model_id: { type: string }
        display_name: { type: string }
        model_type: { type: string, enum: [chat, embedding, rerank] }
        status: { type: string }
        context_length: { type: integer }
        capabilities: {}
        pricing: {}
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        provider:
          $ref: '#/components/schemas/ProviderInfo'